HPCC-26594 React ECL Watch pages not refreshing data
Refreshing the Table doesn't always refresh the data underneath.
Removed superfluous overflowButtonProps={{}}.
Added missing Refresh Buttons.
Split out Summary tabs from Files and Workunits.
