+++ esp/src/package.json
@@ -33,28 +33,28 @@
   },
   "main": "src/stub.js",
   "dependencies": {
-    "@fluentui/react": "^8.34.4",
+    "@fluentui/react": "8.35.0",
     "@fluentui/react-cards": "1.0.0-beta.0",
-    "@fluentui/react-hooks": "^8.3.1",
-    "@fluentui/react-icons-mdl2": "^1.2.1",
-    "@hpcc-js/chart": "^2.67.0",
-    "@hpcc-js/codemirror": "^2.49.0",
-    "@hpcc-js/common": "^2.57.0",
-    "@hpcc-js/comms": "^2.56.0",
-    "@hpcc-js/dataflow": "^3.0.1",
-    "@hpcc-js/eclwatch": "^2.56.0",
-    "@hpcc-js/graph": "^2.69.0",
-    "@hpcc-js/html": "^2.32.0",
-    "@hpcc-js/layout": "^2.35.0",
-    "@hpcc-js/map": "^2.62.0",
-    "@hpcc-js/other": "^2.13.68",
-    "@hpcc-js/phosphor": "^2.14.49",
-    "@hpcc-js/react": "^2.40.0",
-    "@hpcc-js/tree": "^2.30.0",
-    "@hpcc-js/util": "^2.38.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@fluentui/react-hooks": "8.3.2",
+    "@fluentui/react-icons-mdl2": "1.2.2",
+    "@hpcc-js/chart": "2.67.0",
+    "@hpcc-js/codemirror": "2.49.0",
+    "@hpcc-js/common": "2.57.0",
+    "@hpcc-js/comms": "2.57.0",
+    "@hpcc-js/dataflow": "3.0.1",
+    "@hpcc-js/eclwatch": "2.57.0",
+    "@hpcc-js/graph": "2.69.0",
+    "@hpcc-js/html": "2.32.0",
+    "@hpcc-js/layout": "2.35.0",
+    "@hpcc-js/map": "2.62.0",
+    "@hpcc-js/other": "2.13.68",
+    "@hpcc-js/phosphor": "2.14.49",
+    "@hpcc-js/react": "2.40.0",
+    "@hpcc-js/tree": "2.30.0",
+    "@hpcc-js/util": "2.38.0",
+    "@material-ui/core": "4.12.3",
+    "@material-ui/icons": "4.11.2",
+    "@material-ui/lab": "4.0.0-alpha.60",
     "clipboard": "2.0.4",
     "detect-browser": "5.0.0",
     "dijit": "1.16.3",
@@ -64,40 +64,40 @@
     "dojox": "1.16.3",
     "es6-promise": "4.2.8",
     "font-awesome": "4.7.0",
-    "formik": "^2.2.9",
+    "formik": "2.2.9",
     "query-string": "6.13.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.13.1",
-    "react-hook-form": "^7.15.4",
-    "react-reflex": "^4.0.3",
-    "react-sizeme": "^3.0.2",
-    "universal-router": "^9.1.0"
+    "react": "16.14.0",
+    "react-dom": "16.14.0",
+    "react-hook-form": "7.16.1",
+    "react-reflex": "4.0.3",
+    "react-sizeme": "3.0.2",
+    "universal-router": "9.1.0"
   },
   "devDependencies": {
-    "@types/dojo": "^1.9.43",
-    "@types/react": "^16.14.15",
-    "@types/react-dom": "^16.9.14",
-    "@typescript-eslint/eslint-plugin": "^4.31.2",
-    "@typescript-eslint/parser": "^4.31.2",
+    "@types/dojo": "1.9.43",
+    "@types/react": "16.14.15",
+    "@types/react-dom": "16.9.14",
+    "@typescript-eslint/eslint-plugin": "4.32.0",
+    "@typescript-eslint/parser": "4.32.0",
     "braces": ">=2.3.1",
-    "cpx": "^1.5.0",
-    "css-loader": "^3.4.2",
-    "dojo-webpack-plugin": "^2.8.20",
-    "eslint": "^7.32.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "file-loader": "^5.1.0",
-    "local-web-server": "^4.0.0",
+    "cpx": "1.5.0",
+    "css-loader": "3.6.0",
+    "dojo-webpack-plugin": "2.8.20",
+    "eslint": "7.32.0",
+    "eslint-plugin-react-hooks": "4.2.0",
+    "file-loader": "5.1.0",
+    "local-web-server": "4.2.1",
     "minimist": ">=1.2.2",
-    "npm-run-all": "^4.1.5",
-    "rimraf": "^3.0.2",
-    "source-map-loader": "^1.1.3",
-    "style-loader": "^1.1.3",
-    "tslib": "^2.3.1",
-    "typescript": "^4.4.3",
-    "url-loader": "^3.0.0",
-    "webpack": "^4.45.0",
-    "webpack-cli": "^4.8.0",
-    "webpack-dev-server": "^3.11.2"
+    "npm-run-all": "4.1.5",
+    "rimraf": "3.0.2",
+    "source-map-loader": "1.1.3",
+    "style-loader": "1.3.0",
+    "tslib": "2.3.1",
+    "typescript": "4.4.3",
+    "url-loader": "3.0.0",
+    "webpack": "4.46.0",
+    "webpack-cli": "4.8.0",
+    "webpack-dev-server": "3.11.2"
   },
   "author": "HPCC Systems",
   "license": "Apache-2.0",
