+++ esp/src/src-react/components/FileHistory.tsx
@@ -1,17 +1,12 @@
 import * as React from "react";
 import { CommandBar, ContextualMenuItemType, ICommandBarItemProps } from "@fluentui/react";
 import { useConst } from "@fluentui/react-hooks";
-import { scopedLogger } from "@hpcc-js/util";
-import * as Observable from "dojo/store/Observable";
-import { Memory } from "src/Memory";
+import { Memory, Observable } from "src/Memory";
 import nlsHPCC from "src/nlsHPCC";
+import { useFileHistory } from "../hooks/file";
 import { useGrid } from "../hooks/grid";
-import { useFile } from "../hooks/file";
 import { HolyGrail } from "../layouts/HolyGrail";
-import * as WsDfu from "../../src/WsDfu";
 import { ShortVerticalDivider } from "./Common";
-
-const logger = scopedLogger("../components/FileHistory.tsx");
 
 interface FileHistoryProps {
     cluster: string;
@@ -23,38 +18,9 @@ export const FileHistory: React.FunctionComponent<FileHistoryProps> = ({
     logicalFile
 }) => {
 
-    const [file, , _refresh] = useFile(cluster, logicalFile);
-
-    //  Command Bar  ---
-    const buttons: ICommandBarItemProps[] = [
-        {
-            key: "refresh", text: nlsHPCC.Refresh, iconProps: { iconName: "Refresh" },
-            onClick: () => refreshTable()
-        },
-        { key: "divider_1", itemType: ContextualMenuItemType.Divider, onRender: () => <ShortVerticalDivider /> },
-        {
-            key: "erase", text: nlsHPCC.EraseHistory,
-            onClick: () => {
-                if (confirm(nlsHPCC.EraseHistoryQ + "\n" + file?.Name + "?")) {
-                    WsDfu.EraseHistory({
-                        request: {
-                            Name: file?.Name
-                        }
-                    })
-                        .then(response => {
-                            if (response) {
-                                store.setData([]);
-                                refreshTable();
-                            }
-                        })
-                        .catch(logger.error)
-                        ;
-                }
-            }
-        },
-    ];
-
     //  Grid ---
+    const [history, eraseHistory, refreshData] = useFileHistory(cluster, logicalFile);
+
     const store = useConst(new Observable(new Memory("Name")));
     const [Grid, _selection, refreshTable, copyButtons] = useGrid({
         store,
@@ -72,36 +38,31 @@ export const FileHistory: React.FunctionComponent<FileHistoryProps> = ({
     });
 
     React.useEffect(() => {
-        WsDfu.ListHistory({
-            request: {
-                Name: file?.Name
-            }
-        }).then(response => {
-            const results = response?.ListHistoryResponse?.History?.Origin;
+        store.setData(history);
+        refreshTable();
+    }, [history, refreshTable, store]);
 
-            if (results) {
-                store.setData(results.map(row => {
-                    return {
-                        Name: row.Name,
-                        IP: row.IP,
-                        Operation: row.Operation,
-                        Owner: row.Owner,
-                        Path: row.Path,
-                        Timestamp: row.Timestamp,
-                        Workunit: row.Workunit
-                    };
-                }));
-                refreshTable();
+    //  Command Bar  ---
+    const buttons: ICommandBarItemProps[] = React.useMemo(() => [
+        {
+            key: "refresh", text: nlsHPCC.Refresh, iconProps: { iconName: "Refresh" },
+            onClick: () => refreshData()
+        },
+        { key: "divider_1", itemType: ContextualMenuItemType.Divider, onRender: () => <ShortVerticalDivider /> },
+        {
+            key: "erase", text: nlsHPCC.EraseHistory, disabled: history?.length === 0,
+            onClick: () => {
+                if (confirm(nlsHPCC.EraseHistoryQ + "\n" + logicalFile + "?")) {
+                    eraseHistory();
+                }
             }
-        }).catch(logger.error)
-            ;
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [store, file?.Name]);
+        },
+    ], [eraseHistory, history?.length, logicalFile, refreshData]);
 
     return <HolyGrail
         header={<CommandBar items={buttons} farItems={copyButtons} />}
         main={
-            < Grid />
+            <Grid />
         }
     />;
 };
