+++ esp/src/src-react/components/ProtectedBy.tsx
@@ -1,15 +1,12 @@
 import * as React from "react";
+import { CommandBar, ICommandBarItemProps } from "@fluentui/react";
 import { useConst } from "@fluentui/react-hooks";
-import { scopedLogger } from "@hpcc-js/util";
 import * as Observable from "dojo/store/Observable";
 import { Memory } from "src/Memory";
 import nlsHPCC from "src/nlsHPCC";
 import { useGrid } from "../hooks/grid";
 import { useFile } from "../hooks/file";
 import { HolyGrail } from "../layouts/HolyGrail";
-import * as WsDfu from "../../src/WsDfu";
-
-const logger = scopedLogger("../components/ProtectedBy.tsx");
 
 interface ProtectedByProps {
     cluster: string;
@@ -21,11 +18,11 @@ export const ProtectedBy: React.FunctionComponent<ProtectedByProps> = ({
     logicalFile
 }) => {
 
-    const [file, , _refresh] = useFile(cluster, logicalFile);
+    const [file, , , refreshData] = useFile(cluster, logicalFile);
 
     //  Grid ---
     const store = useConst(new Observable(new Memory("Owner")));
-    const [Grid, _selection, refreshTable, _copyButtons] = useGrid({
+    const [Grid, _selection, refreshTable, copyButtons] = useGrid({
         store,
         sort: [{ attribute: "Owner", "descending": false }],
         filename: "protectedBy",
@@ -36,28 +33,29 @@ export const ProtectedBy: React.FunctionComponent<ProtectedByProps> = ({
     });
 
     React.useEffect(() => {
-        WsDfu.DFUInfo({
-            request: {
-                Name: file?.Name
-            }
-        }).then(response => {
-            const results = response?.DFUInfoResponse?.FileDetail.ProtectList.DFUFileProtect;
+        const results = file?.ProtectList?.DFUFileProtect;
 
-            if (results) {
-                store.setData(results.map(row => {
-                    return {
-                        Owner: row.Owner,
-                        Modified: row.Modified
-                    };
-                }));
-                refreshTable();
-            }
-        })
-            .catch(logger.error)
-            ;
-    }, [store, file?.Name, refreshTable]);
+        if (results) {
+            store.setData(file?.ProtectList?.DFUFileProtect?.map(row => {
+                return {
+                    Owner: row.Owner,
+                    Modified: row.Modified
+                };
+            }));
+            refreshTable();
+        }
+    }, [store, file?.ProtectList?.DFUFileProtect, refreshTable]);
+
+    //  Command Bar  ---
+    const buttons = React.useMemo((): ICommandBarItemProps[] => [
+        {
+            key: "refresh", text: nlsHPCC.Refresh, iconProps: { iconName: "Refresh" },
+            onClick: () => refreshData()
+        },
+    ], [refreshData]);
 
     return <HolyGrail
+        header={<CommandBar items={buttons} farItems={copyButtons} />}
         main={
             <Grid />
         }
